+++ lib/src/ast/sass/argument.dart
@@ -7,13 +7,14 @@ import 'package:source_span/source_span.dart';
 
 import '../../utils.dart';
 import 'expression.dart';
+import 'interface/declaration.dart';
 import 'node.dart';
 
 /// An argument declared as part of an [ArgumentDeclaration].
 ///
 /// {@category AST}
 @sealed
-class Argument implements SassNode {
+class Argument implements SassNode, SassDeclaration {
   /// The argument name.
   final String name;
 
@@ -29,6 +30,9 @@ class Argument implements SassNode {
   /// messages.
   String get originalName =>
       defaultValue == null ? span.text : declarationName(span);
+
+  FileSpan get nameSpan =>
+      defaultValue == null ? span : span.subspan(0, name.length + 1);
 
   Argument(this.name, this.span, {this.defaultValue});