+++ lib/src/ast/sass/expression/variable.dart
@@ -7,12 +7,13 @@ import 'package:source_span/source_span.dart';
 
 import '../../../visitor/interface/expression.dart';
 import '../expression.dart';
+import '../interface/reference.dart';
 
 /// A Sass variable.
 ///
 /// {@category AST}
 @sealed
-class VariableExpression implements Expression {
+class VariableExpression implements Expression, SassReference {
   /// The namespace of the variable being referenced, or `null` if it's
   /// referenced without a namespace.
   final String? namespace;
@@ -21,6 +22,13 @@ class VariableExpression implements Expression {
   final String name;
 
   final FileSpan span;
+
+  FileSpan get nameSpan =>
+      namespace == null ? span : span.subspan(namespace!.length + 1);
+
+  FileSpan get namespaceSpan => namespace == null
+      ? span.start.pointSpan()
+      : span.subspan(0, namespace!.length);
 
   VariableExpression(this.name, this.span, {this.namespace});