+++ lib/src/ast/sass/statement/mixin_rule.dart
@@ -8,6 +8,7 @@ import 'package:source_span/source_span.dart';
 import '../../../visitor/interface/statement.dart';
 import '../../../visitor/statement_search.dart';
 import '../argument_declaration.dart';
+import '../interface/declaration.dart';
 import '../statement.dart';
 import 'callable_declaration.dart';
 import 'content_rule.dart';
@@ -19,10 +20,16 @@ import 'silent_comment.dart';
 ///
 /// {@category AST}
 @sealed
-class MixinRule extends CallableDeclaration {
+class MixinRule extends CallableDeclaration implements SassDeclaration {
   /// Whether the mixin contains a `@content` rule.
   late final bool hasContent =
       const _HasContentVisitor().visitMixinRule(this) == true;
+
+  FileSpan get nameSpan {
+    var match = RegExp(r'(=|@mixin)\s*').matchAsPrefix(span.text);
+    var start = match!.end;
+    return span.subspan(start, start + name.length);
+  }
 
   MixinRule(String name, ArgumentDeclaration arguments,
       Iterable<Statement> children, FileSpan span,