+++ pkg/sass_api/CHANGELOG.md
@@ -1,3 +1,20 @@
+## 1.0.0-beta.4
+
+* `UseRule`, `ForwardRule`, and `DynamicImport` now share a common `Dependency`
+  interface that exposes a `url` getter and a `urlSpan` getter.
+
+* `VariableDeclaration`, `CallableDeclaration`, `Argument`, and
+  `ConfiguredVariable` now share a common `SassDeclaration` interface that
+  exposes a `name` getter (with underscores converted to hyphens) and a
+  `nameSpan` getter.
+
+* `VariableExpression`, `IncludeRule`, and `FunctionExpression` now share a
+  common `SassReference` interface that exposes a `namespace` getter and a
+  `name` getter (which has underscores converted to hyphens, or is null for
+  `FunctionExpression`s with interpolation), as well as corresponding
+  `namespaceSpan` and `nameSpan` getters. The old `FunctionExpression.name`
+  getter has been renamed to `FunctionExpression.interpolatedName`.
+
 ## 1.0.0-beta.3
 
 * No user-visible changes.